Extract shared carpooling service test wiring
